Require explicit development mode
Before, anything other than `ENV=production` would set the logger in
development mode. This is somewhat "dangerous", as it means that if
you don't set this variable correctly, Zap's log sampler is disabled,
potentially causing millions of lines to be logged, causing an
application to crawl to a halt.

With this change, only `ENV=development` sets this flag. You can
still enable this manually:

    logger.New("my-app", "v1.0.0", zap.Development())